requirements.txt
django==3.2.15
psycopg2==2.9.3
djangorestframework==3.14.0
